+++ lib/rails_agents.rb
@@ -7,8 +7,12 @@ require "rails_agents/credentials_writer"
 require "rails_agents/local_auth"
 require "rails_agents/local_sync"
 require "rails_agents/base"
-require "rails_agents/chat_agent"
+require "rails_agents/knowledge_agent"
+require "rails_agents/creation_agent"
 require "rails_agents/workflow_agent"
+require "rails_agents/operations_agent"
+require "rails_agents/monitoring_agent"
+require "rails_agents/chat_agent"
 require "rails_agents/background_agent"
 require "rails_agents/cli"
 
@@ -21,6 +25,9 @@ if defined?(Rails::Railtie)
 end
 
 module RailsAgents
+  TAXONOMY_TYPES = %i[knowledge workflow operations monitoring].freeze
+  LEGACY_TAXONOMY_TYPES = %i[creation].freeze
+
   class << self
     attr_writer :config